Handle Errors/Response for User
Handle Errors for User:

Show user if is Ok or some Error happened, for this I use the response after submit to show everighing is good.
Else user will see the message error for this I take ventage from error send by the APi.

Also a new method to reset the message in screen by passing in args the time.