EventEmitter
Each time when a record is push to DataBase, the succes response emit a signal to Table to refresh all, and render the new input.
Service has an action emiter (string), and two emiter's in Calc Componet, how emit the signal to History comonent where's suscription to this an get signal to make new call to DB for the last insert.